+++ quadratic-core/examples/v1_4_full_featured.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.5",
+  "version": "1.4",
   "sheets": [
     {
       "id": { "id": "79f556ac-13c8-41b5-b407-55c0a1435022" },
