+++ quadratic-core/examples/v1_4_simple.json
@@ -212,5 +212,5 @@
       ]
     }
   ],
-  "version": "1.5"
+  "version": "1.4"
 }
