+++ src/schemas/validateAndUpgradeLegacyGridFile.ts
@@ -13,7 +13,7 @@ const FILES = [
 ];
 
 // all versions validated by Rust
-const rustFileVersions = ['1.3', '1.5'];
+const rustFileVersions = ['1.3', '1.4'];
 
 /**
  * Given arbitrary JSON, validate whether it's a valid file format and return
